+++ android/apollo/src/main/java/io/muun/apollo/domain/model/P2PState.kt
@@ -5,7 +5,7 @@ import io.muun.apollo.domain.model.user.User
 
 class P2PState(
     val user: User,
-    val permissionState: ContactsPermissionState,
+    val permissionState: PermissionState,
     val syncState: ActionState<*>,
     val contacts: List<Contact>
 )