+++ android/apollo/src/main/java/io/muun/apollo/domain/model/PermissionState.java
@@ -1,7 +1,7 @@
 package io.muun.apollo.domain.model;
 
-public enum ContactsPermissionState {
-
+public enum PermissionState {
+    NOT_DETERMINED,
     GRANTED,
     DENIED,
     PERMANENTLY_DENIED //aka Never Ask Again